refactor(client): HASHI-80 식당 리스트 기능 구조 이동
