+++ apps/client/src/features/restaurantList/RestaurantImageList.tsx
