+++ apps/client/src/features/restaurantList/constants.ts
